Implement the add() method
I decided to implement the add() method first in order to make sure I was building the tree correctly before continuing with the project. This commit also includes a 'printPreOrder()' function that I used to test the add method. In addition I created a helper method for the add method called comparePrefix() that returns the index of the first letter in both the word and the prefix's suffix (if it exists.)