Implement check() method
This commit includes the completed check() method that finds words in the tree. The method is similar to the add() method in many ways, but requires less steps. The method checks for the base cases and if none of them are satisfied it recursively checks the child index where the suffix of the word should be stored. I tweaked the test code a little bit to check for all of the words that were added and to print them back to the console if the method returns true. It found every single word and successfully ignored case.