Fixed bug in check() method
Realized that I had implemented the check() method incorrectly and that my test code was also incorrect. I hadn't tested it with a word that was not valid yet, and as soon as I did it broke. I had made the check() method too similar to the add() method and didn't take into consideration that in find() the whole prefix had to be part of the word that was being searched for. I had originally written it in the way it's written in the add method, where it just checks for any common prefix between the node and the word being passed. I fixed these issues and tested them thoroughly.