Implemented the toString() Method
Created a private toString() method to recursively create a string that contains the human readable tree. I call the private method from the public method by inserting the root and the numIndents as 0. Implementing this method made me realize a bug in the add() method that was not updating new suffixes being added to be considered words. I fixed that in this commit as well.