Implemented the Suggest() Method
I implemented and tested a suggest() method for the prefix tree. I used two helper methods which gather all of the suggestions into an array list that is then convereted back in to a string array and returned in the public suggest() method. When all of the suggestions for the input are exhausted, the code begins chopping off letters from the input and offering suggestions based on similar prefixes. The code is passing all of the tests provided by the instructor as well as some written by me.